fix: isolate protocol version features (#19)
* fix: isolate protocol version features

* fix: adapt unastar to isolated protocol types

* fix: link Windows LevelDB against static zlib

+++ crates/unastar/src/server/game/commands.rs
@@ -271,7 +271,7 @@ fn handle_teleport_command(
 
     // Send teleport packet
     let packet = MovePlayerPacket {
-        runtime_id: runtime_id as i32,
+        runtime_id: runtime_id as u64,
         position: Vec3F {
             x: x as f32,
             y: y as f32,
